Update hate_crack.py for Windows compatibility
Added module codecs and switched from iso-8859-9 to utf-8 to support and display umlauts on Windows correctly.

Added if/elif clause and some variables e.g. for quotes and directory separator for Windows compatibility.

Modified hcatMiddleCombinator() and hcatThoroughCombinator() to support multiple character masks on Windows.

Added support for multiple single quote characters as a mask on Linux.

Modified quick_crack() to avoid a syntax warning message on Python 3.8+ when identity checks (is and is not) are used with certain types of literals (e.g. strings, numbers).